backend/requirements.txt
# FastAPI core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Authentication
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Validation & Settings
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# AI - Doubao (Volcengine)
volcengine-python-sdk[ark]==1.0.0

# PDF Generation
weasyprint==62.3

# Utilities
httpx==0.27.2
python-dateutil==2.9.0
Pillow==10.4.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
